--- src/_ann_module3.py
@@ -1,18 +1,0 @@
-"""
-Correct syntax for variable annotation that should fail at runtime in a certain manner.
-This file is for testing purposes, and is vendored from Lib/test/ann_module3.py in CPython.
-"""
-
-def f_bad_ann():
-    __annotations__[1] = 2
-
-class C_OK:
-    def __init__(self, x: int) -> None:
-        self.x: no_such_name = x  # This one is OK as proposed by Guido
-
-class D_bad_ann:
-    def __init__(self, x: int) -> None:
-        sfel.y: int = 0
-
-def g_bad_ann():
-    no_such_name.attr: int = 0
